Fixed a few warnings.
Rev: src/builtin_functions.c:1.298
Rev: src/docode.c:1.79
Rev: src/encode.c:1.67
Rev: src/gc.c:1.122
Rev: src/las.c:1.191
Rev: src/module_support.c:1.36
Rev: src/peep.c:1.35
Rev: src/pike_memory.h:1.22
Rev: src/pike_types.c:1.135
Rev: src/pike_types.h:1.44
Rev: src/program.c:1.262

+++ src/pike_types.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: pike_types.h,v 1.43 2000/08/15 15:59:49 grubba Exp $
+ * $Id: pike_types.h,v 1.44 2000/08/16 10:26:51 grubba Exp $
  */
 #ifndef PIKE_TYPES_H
 #define PIKE_TYPES_H
@@ -103,8 +103,8 @@ extern struct pike_string *weak_type_string;
 #ifdef PIKE_DEBUG
 #define init_type_stack() type_stack_mark()
 #define exit_type_stack() do {\
-  int q_q_q_q=pop_stack_mark(); \
-  if(q_q_q_q) fatal("Type stack out of wack! %d\n",q_q_q_q); \
+    ptrdiff_t q_q_q_q = pop_stack_mark(); \
+    if(q_q_q_q) fatal("Type stack out of wack! %d\n", q_q_q_q); \
   } while(0)
 #else
 #define init_type_stack type_stack_mark
